requirements.txt
streamlit==1.44.0
langchain==0.3.21
langgraph==0.3.21
langchain-core==0.3.49
langchain-community==0.3.20
langchain-mcp-adapters==0.0.5
langchain-teddynote==0.3.45
mcp==1.4.1
nest-asyncio==1.6.0
python-dotenv==1.1.0
anyio==4.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
deepl==1.21.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.0
faiss-cpu==1.10.0
fastjsonschema==2.21.1
feedparser==6.0.11
fqdn==1.5.1
frozenlist==1.5.0
gitdb==4.0.12
gitpython==3.1.44
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kiwipiepy==0.20.4
kiwipiepy-model==0.20.0
langchain-anthropic==0.3.10
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.19
lz4==4.4.3
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.1.3
mmh3==4.1.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nltk==3.9.1
notebook==7.3.3
notebook-shim==0.2.4
numpy==1.26.4
olefile==0.47
openai==1.69.0
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.1.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.10.0
platformdirs==4.3.7
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
protoc-gen-openapiv2==0.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.0
pydantic-core==2.33.0
pydantic-settings==2.8.1
pydeck==0.9.1
pygments==2.19.1
pymupdf==1.25.4
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.3.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
send2trash==1.8.3
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.40
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
tavily-python==0.5.2
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250328
typing-extensions==4.13.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.0
zstandard==0.23.0
traitlets==5.14.3
types-requests==2.32.0.20250328
typing-extensions==4.13.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xxhash==3.5.0
zstandard==0.23.0

# 구글 api 관련
google-api-python-client
google_auth_oauthlib

# 솔라 api 관련
langchain-upstage
